Add Azure DocumentDB integration
