+++ CONTRIBUTING.md
@@ -33,10 +33,11 @@ Install uv: **[documentation on how to install it](https://docs.astral.sh/uv/get
 
 ## Different packages
 
-This repository contains seven packages with Azure integrations with LangChain:
+This repository contains eight packages with Azure integrations with LangChain:
 - [langchain-azure-ai](https://pypi.org/project/langchain-azure-ai/)
 - [langchain-azure-compute](./libs/azure-compute)
 - [langchain-azure-cosmosdb](https://pypi.org/project/langchain-azure-cosmosdb/)
+- [langchain-azure-documentdb](./libs/azure-documentdb)
 - [langchain-azure-dynamic-sessions](https://pypi.org/project/langchain-azure-dynamic-sessions/)
 - [langchain-sqlserver](https://pypi.org/project/langchain-sqlserver/)
 - [langchain-azure-postgresql](https://pypi.org/project/langchain-azure-postgresql/)
@@ -60,6 +61,7 @@ Here's the structure visualized as a tree:
 │   ├── azure-ai
 │   ├── azure-compute
 │   ├── azure-cosmosdb
+│   ├── azure-documentdb
 │   ├── azure-dynamic-sessions
 │   ├── azure-postgresql
 │   ├── azure-storage
