+++ libs/azure-ai/langchain_azure_ai/vectorstores/__init__.py
@@ -34,7 +34,7 @@ __all__ = [
 ]
 
 _module_lookup = {
-    "AzureDocumentDBVectorSearch": "langchain_azure_ai.vectorstores.azure_cosmos_db_mongo_vcore",  # noqa: E501
+    "AzureDocumentDBVectorSearch": "langchain_azure_ai.vectorstores.azure_documentdb",  # noqa: E501
     "AzureCosmosDBMongoVCoreVectorSearch": "langchain_azure_ai.vectorstores.azure_cosmos_db_mongo_vcore",  # noqa: E501
     "AzureCosmosDBNoSqlVectorSearch": "langchain_azure_ai.vectorstores.azure_cosmos_db_no_sql",  # noqa: E501
     "AzureSearch": "langchain_azure_ai.vectorstores.azuresearch",
