+++ libs/azure-ai/langchain_azure_ai/vectorstores/azure_cosmos_db_mongo_vcore.py
@@ -1,10 +1,10 @@
-"""Vector Store for Azure DocumentDB (with MongoDB compatibility) — DEPRECATED.
+"""Legacy compatibility imports for the renamed Azure DocumentDB integration.
 
-This module has moved to ``langchain_azure_cosmosdb``.
+This module has moved to ``langchain_azure_documentdb``.
 Install and import directly from there instead::
 
-    pip install langchain-azure-cosmosdb
-    from langchain_azure_cosmosdb import AzureDocumentDBVectorSearch
+    pip install langchain-azure-documentdb
+    from langchain_azure_documentdb import AzureDocumentDBVectorSearch
 """
 
 import warnings
@@ -18,42 +18,35 @@ _DEPRECATED_NAMES = {
     "CosmosDBVectorSearchType",
 }
 
+_DOCUMENTDB_NAMES = {
+    "AzureDocumentDBVectorSearch": "AzureDocumentDBVectorSearch",
+    "AzureCosmosDBMongoVCoreVectorSearch": "AzureDocumentDBVectorSearch",
+    "CosmosDBSimilarityType": "AzureDocumentDBSimilarityType",
+    "CosmosDBVectorSearchCompression": "AzureDocumentDBVectorSearchCompression",
+    "CosmosDBVectorSearchType": "AzureDocumentDBVectorSearchType",
+}
+
 
 def __getattr__(name: str) -> Any:
     if name in _DEPRECATED_NAMES:
+        documentdb_name = _DOCUMENTDB_NAMES[name]
         warnings.warn(
             f"Importing {name} from "
             "'langchain_azure_ai.vectorstores.azure_cosmos_db_mongo_vcore' is "
             "deprecated. "
-            f"Use 'from langchain_azure_cosmosdb import {name}' instead.",
+            "Use 'from langchain_azure_documentdb import "
+            f"{documentdb_name}' instead.",
             DeprecationWarning,
             stacklevel=2,
         )
         try:
-            import langchain_azure_cosmosdb
+            import langchain_azure_documentdb
 
-            _map: dict[str, Any] = {
-                "AzureDocumentDBVectorSearch": (
-                    langchain_azure_cosmosdb.AzureDocumentDBVectorSearch
-                ),
-                "AzureCosmosDBMongoVCoreVectorSearch": (
-                    langchain_azure_cosmosdb.AzureCosmosDBMongoVCoreVectorSearch
-                ),
-                "CosmosDBSimilarityType": (
-                    langchain_azure_cosmosdb.CosmosDBSimilarityType
-                ),
-                "CosmosDBVectorSearchCompression": (
-                    langchain_azure_cosmosdb.CosmosDBVectorSearchCompression
-                ),
-                "CosmosDBVectorSearchType": (
-                    langchain_azure_cosmosdb.CosmosDBVectorSearchType
-                ),
-            }
-            return _map[name]
-        except ImportError:
+            return getattr(langchain_azure_documentdb, documentdb_name)
+        except ImportError as exc:
             raise ImportError(
-                f"langchain-azure-cosmosdb is required for {name}. "
-                "Install it with: pip install langchain-azure-cosmosdb"
-            )
+                f"langchain-azure-documentdb is required for {name}. "
+                "Install it with: pip install langchain-azure-documentdb"
+            ) from exc
 
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
