+++ libs/azure-ai/pyproject.toml
@@ -37,6 +37,7 @@ dependencies = [
   "azure-identity~=1.15",
   "azure-search-documents>=11.4,<13.0",
   "langchain>=1.2.12,<2.0.0",
+  "langchain-azure-documentdb>=0.1.0,<1.0.0",
   "langchain-openai>=1.0.0,<2.0.0",
   "langgraph>=1.1.1,<3.0",
   "numpy>=1.26.2; python_version < '3.13'",
@@ -202,3 +203,4 @@ typing = ["mypy>=1.10,<3.0", "types-requests~=2.28"]
 
 [tool.uv.sources]
 langchain-core = { git = "https://github.com/langchain-ai/langchain.git", subdirectory = "libs/core" }
+langchain-azure-documentdb = { path = "../azure-documentdb", editable = true }
