+++ libs/azure-documentdb/.gitignore
@@ -1,0 +1,3 @@
+__pycache__
+.ruff_cache
+.pytest_cache
