+++ libs/azure-documentdb/pyproject.toml
@@ -1,0 +1,92 @@
+[project]
+name = "langchain-azure-documentdb"
+version = "0.1.0"
+description = "Azure DocumentDB integrations for LangChain, including vector search."
+license = "MIT"
+readme = "README.md"
+requires-python = ">=3.10,<4.0"
+authors = [
+  { name = "Microsoft Corporation" },
+]
+keywords = [
+  "azure",
+  "documentdb",
+  "langchain",
+  "vector-search",
+  "llm",
+  "ai",
+]
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Developers",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Database",
+  "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
+
+dependencies = [
+  "langchain-core>=1.0,<2.0",
+  "numpy>=1,<3",
+  "pymongo>=4.5,<5.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/langchain-ai/langchain-azure/tree/main/libs/azure-documentdb"
+repository = "https://github.com/langchain-ai/langchain-azure"
+"Source Code" = "https://github.com/langchain-ai/langchain-azure/tree/main/libs/azure-documentdb"
+"Release Notes" = "https://github.com/langchain-ai/langchain-azure/releases"
+
+[build-system]
+requires = ["uv_build>=0.8,<0.13"]
+build-backend = "uv_build"
+
+[tool.uv.build-backend]
+module-name = "langchain_azure_documentdb"
+module-root = "src"
+
+[tool.coverage.run]
+omit = ["tests/*"]
+
+[tool.mypy]
+mypy_path = "src"
+disallow_untyped_defs = true
+explicit_package_bases = true
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+addopts = [
+  "--durations=5",
+  "--strict-config",
+  "--strict-markers",
+]
+markers = [
+  "compile: mark placeholder test used to compile integration tests without running them",
+  "requires: mark tests as requiring a specific library or Azure service",
+]
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "D"]
+
+[tool.ruff.lint.per-file-ignores]
+"tests/**" = ["D"]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
+
+[tool.uv]
+default-groups = ["typing"]
+
+[dependency-groups]
+codespell = ["codespell~=2.2"]
+lint = ["pytest>=9.0.3,<10", "ruff>=0.5,<0.17"]
+test = ["pytest>=9.0.3,<10"]
+test_integration = [
+  "azure-identity>=1.25,<2.0",
+  "langchain-openai>=1.0,<2.0",
+  "pytest>=9.0.3,<10",
+]
+typing = ["mypy>=1.10,<3.0"]
