+++ libs/azure-documentdb/src/langchain_azure_documentdb/__init__.py
@@ -1,0 +1,15 @@
+"""Azure DocumentDB integrations for LangChain."""
+
+from langchain_azure_documentdb._vectorstore import (
+    AzureDocumentDBSimilarityType,
+    AzureDocumentDBVectorSearch,
+    AzureDocumentDBVectorSearchCompression,
+    AzureDocumentDBVectorSearchType,
+)
+
+__all__ = [
+    "AzureDocumentDBSimilarityType",
+    "AzureDocumentDBVectorSearch",
+    "AzureDocumentDBVectorSearchCompression",
+    "AzureDocumentDBVectorSearchType",
+]
