+++ libs/azure-documentdb/tests/integration_tests/test_placeholder.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.mark.compile
+def test_placeholder() -> None:
+    """Compile integration tests without running external services."""
+    pass
